Intent testing for logout, todaySwitch
