+++ app/src/androidTest/java/com/example/routines/LogoutTest.java
@@ -1,0 +1,60 @@
+package com.example.routines;
+
+import android.app.Activity;
+import android.view.View;
+import android.widget.EditText;
+
+import androidx.test.platform.app.InstrumentationRegistry;
+import androidx.test.rule.ActivityTestRule;
+
+import com.robotium.solo.Solo;
+
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+
+public class LogoutTest {
+    private Solo solo;
+    @Rule // start testing on welcome page
+    public ActivityTestRule<WelcomeActivity> rule =
+            new ActivityTestRule<>(WelcomeActivity.class, true, true);
+
+    /**
+     * Runs before all tests and creates solo instances
+     * @throws Exception
+     */
+    @Before
+    public void setup() throws Exception{
+        solo = new Solo(InstrumentationRegistry.getInstrumentation(), rule.getActivity());
+    }
+    public void start() throws Exception{
+        Activity activity = rule.getActivity();
+    }
+
+    public void login(){
+        solo.assertCurrentActivity("Wrong activity, needs to be Welcome", WelcomeActivity.class);
+        solo.clickOnButton("LOGIN");
+        solo.assertCurrentActivity("Activity needs to be loginActivity", LoginActivity.class);
+        solo.enterText((EditText) solo.getView(R.id.editText_email), "testEmail@gmail.com");
+        solo.enterText((EditText) solo.getView(R.id.editText_password_login), "testPassword123456");
+        solo.clickOnButton("LOGIN");
+        solo.assertCurrentActivity("Activity needs to be homeActivity", HomeActivity.class);
+    }
+
+    @Test
+    public void logoutTest(){
+        login(); // login
+        solo.sleep(500);
+        // get the profile button from the bottom menu
+        View bottomBar= solo.getCurrentActivity().findViewById(R.id.bottom_navigation); // get the button inside the frame layout
+        View profile = bottomBar.findViewById(R.id.profile);
+        solo.clickOnView(profile);
+        solo.sleep(500);
+        // logout
+        solo.assertCurrentActivity("Activity needs to be profile activity", ProfileActivity.class);
+        solo.clickOnButton("LOG OUT");
+        // after logging out we should be back at welcome activity
+        solo.assertCurrentActivity("Activity needs to be welcome activity", WelcomeActivity.class);
+        solo.sleep(500);
+    }
+}
